Create a Parser class for more OOP
